requirements.txt
openai
python-dotenv
discord
pynacl
numpy
pandas
langchain
langchain-core
langchain-community
langchain-experimental
langgraph
langchain-cli
langsmith
scipy
tenacity
tiktoken
termcolor
discord.py[voice]
wave
pyaudio
SpeechRecognition